feat(tee): custom attestation module
Fixes #613
