+++ libs/chain-signatures/Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 members = [
-    "contract",
+    "contract", "tee",
 ]
 resolver = "2"
