+++ libs/chain-signatures/tee/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "tee"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = "1.0.98"
+hex = "0.4.3"
+serde = { version = "1.0.219" }
+serde_json = { version = "1.0.140" }
+dstack-sdk = { git = "https://github.com/Dstack-TEE/dstack.git", tag = "v0.5.2" }
+mpc-contract = { path = "../contract/", features = [
+    "dev-utils",
+] }
