+++ node/Cargo.toml
@@ -65,11 +65,12 @@ near-config-utils = { git = "https://github.com/near/nearcore", tag = "2.7.0-rc.
 near-crypto = { git = "https://github.com/near/nearcore", tag = "2.7.0-rc.2" }
 near-o11y = { git = "https://github.com/near/nearcore", tag = "2.7.0-rc.2" }
 near-time = { git = "https://github.com/near/nearcore", tag = "2.7.0-rc.2" }
-# todo: update once 1.1.0 is pulbished (though the api did not change)
+# todo: update once 1.1.0 is published (though the api did not change)
 legacy-mpc-contract = { package = "mpc-contract", git = "https://github.com/near/mpc/", rev = "1d4954dff28e8eb988fb7762eff414a602a2b124" }
 mpc-contract = { path = "../libs/chain-signatures/contract/", features = [
     "dev-utils",
 ] }
+tee = { path = "../libs/chain-signatures/tee/" }
 # Pinned to resolve compilation errors with conflicting implementations of PartialOrd trait
 # See error E0283 with multiple impls satisfying `u128: std::cmp::PartialOrd<_>`
 deranged = "=0.4.0"
